+++ packages/deployed-backend-client/src/deployed_backend_client.ts
@@ -3,8 +3,10 @@ import {
   UniqueBackendIdentifier,
 } from '@aws-amplify/plugin-types';
 import {
+  ApiAuthType,
   BackendDeploymentStatus,
   BackendMetadata,
+  ConflictResolutionMode,
   DeployedBackendClient,
   ListSandboxesRequest,
   ListSandboxesResponse,
@@ -28,6 +30,9 @@ import {
   StackStatus,
   StackSummary,
 } from '@aws-sdk/client-cloudformation';
+
+import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
+
 import {
   authOutputKey,
   graphqlOutputKey,
@@ -44,6 +49,7 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
    */
   constructor(
     private readonly cfnClient: CloudFormationClient,
+    private readonly s3Client: S3Client,
     private readonly backendOutputClient: BackendOutputClient
   ) {}
 
@@ -217,15 +223,58 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     }
 
     if (apiStack) {
+      const additionalAuthTypesString =
+        backendOutput[graphqlOutputKey]?.payload
+          .awsAppsyncAdditionalAuthenticationTypes;
+      const additionalAuthTypes = additionalAuthTypesString
+        ? (additionalAuthTypesString.split(',') as ApiAuthType[])
+        : [];
       backendMetadataObject.apiConfiguration = {
         status: this.translateStackStatus(apiStack.StackStatus),
         lastUpdated: apiStack.LastUpdatedTime,
         graphqlEndpoint: backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncApiEndpoint as string,
+        defaultAuthType: backendOutput[graphqlOutputKey]?.payload
+          .awsAppsyncAuthenticationType as ApiAuthType,
+        additionalAuthTypes,
+        conflictResolutionMode: backendOutput[graphqlOutputKey]?.payload
+          .awsAppsyncConflictResolutionMode as ConflictResolutionMode,
+        graphqlSchema: await this.fetchGraphqlSchema(
+          backendOutput[graphqlOutputKey]?.payload.amplifyApiModelSchemaS3Uri
+        ),
       };
     }
 
     return backendMetadataObject;
+  };
+
+  private fetchGraphqlSchema = async (
+    graphqlSchemaS3Uri: string | undefined
+  ): Promise<string> => {
+    if (!graphqlSchemaS3Uri) {
+      throw new Error('graphqlSchemaS3Uri output is not available');
+    }
+
+    // s3://{bucketName}/{fileName}
+    const uriParts = graphqlSchemaS3Uri.split('/');
+    const bucketName = uriParts[2];
+    const objectPath = uriParts.slice(3, uriParts.length).join('/');
+
+    if (!bucketName || !objectPath) {
+      throw new Error('graphqlSchemaS3Uri is not valid');
+    }
+
+    const s3Response = await this.s3Client.send(
+      new GetObjectCommand({ Bucket: bucketName, Key: objectPath })
+    );
+
+    if (!s3Response.Body) {
+      throw new Error(
+        `s3Response from ${graphqlSchemaS3Uri} does not contain a Body`
+      );
+    }
+
+    return await s3Response.Body?.transformToString();
   };
 
   private translateStackStatus = (