+++ packages/client-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/client-config
 
+## 0.2.0-alpha.9
+
+### Patch Changes
+
+- 2525b582: add reader for cms metadata
+- Updated dependencies [2525b582]
+  - @aws-amplify/deployed-backend-client@0.2.0-alpha.3
+  - @aws-amplify/backend-output-schemas@0.2.0-alpha.6
+
+## 0.2.0-alpha.8
+
+### Patch Changes
+
+- 36d93e46: add license to package.json
+- Updated dependencies [36d93e46]
+  - @aws-amplify/deployed-backend-client@0.2.0-alpha.2
+  - @aws-amplify/backend-output-schemas@0.2.0-alpha.5
+
 ## 0.2.0-alpha.7
 
 ### Patch Changes